requirements.txt
django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-filter==21.1
djongo==1.3.6
pymongo==3.12.1
dnspython==2.1.0
#psycopg2-binary==2.9.1
gunicorn==20.1.0
django-heroku==0.3.1
